requirements.txt
SpeechRecognition
PyAudio
google-genai
gTTS
pygame